requirements.txt
cycler==0.10.0
decorator==4.4.2
kiwisolver==1.2.0
matplotlib==3.2.1
networkx==2.4
numpy==1.18.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
six==1.14.0
